requirements.txt
# Trader & FastAPI Backend
finance-datareader==0.9.90
# Install Meta's Prophet. if os is linux install fbprophet else prophet
# fbprophet==1.1.5
prophet==1.1.5
yfinance==0.2.37
statsmodels==0.14.1
fastapi==0.100
uvicorn==0.29.0
inject==4.3.1
sqlmodel==0.0.9
sqlalchemy==1.4.49
# Airflow
python-dotenv==1.0.1
